Moved argument parsing into its own class and began adding better argument support to PathCheckerGUI
