+++ src/PathLengthChecker/PathLengthChecker.csproj
@@ -47,6 +47,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ArgumentParser.cs" />
     <Compile Include="FileSystemTypes.cs" />
     <Compile Include="MinPathLengthGreaterThanMaxPathLengthException.cs" />
     <Compile Include="PathInfo.cs" />
